# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: 99Placement Recruitment OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /screening:
    post:
      operationId: createScreening
      tags: [screening]
      summary: Create a screening interview record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScreeningRequest"
      responses:
        "201":
          description: Created screening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
  /screening/{id}:
    get:
      operationId: getScreening
      tags: [screening]
      summary: Get screening with criteria scores
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Screening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
    patch:
      operationId: updateScreening
      tags: [screening]
      summary: Save/update screening scores
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScreeningRequest"
      responses:
        "200":
          description: Updated screening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
  /screening/{id}/submit:
    post:
      operationId: submitScreening
      tags: [screening]
      summary: Lock screening and set verdict
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVerdictRequest"
      responses:
        "200":
          description: Submitted screening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"

  # ── ASSESSMENTS ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ScreeningCriteriaScore:
      type: object
      properties:
        criterion:
          type: string
        score:
          type: integer
        notes:
          type: string
          nullable: true
      required: [criterion, score]
    Screening:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        interviewerId:
          type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        conductedAt:
          type: string
          format: date-time
          nullable: true
        durationMin:
          type: integer
          nullable: true
        mode:
          type: string
        overallScore:
          type: integer
          nullable: true
        verdict:
          type: string
          enum: [shortlist, hold, reject]
          nullable: true
        notes:
          type: string
          nullable: true
        criteriaScores:
          type: array
          items:
            $ref: "#/components/schemas/ScreeningCriteriaScore"
        createdAt:
          type: string
          format: date-time
      required: [id, pipelineId, interviewerId, mode, createdAt, criteriaScores]
    CreateScreeningRequest:
      type: object
      properties:
        pipelineId:
          type: string
        interviewerId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        mode:
          type: string
        notes:
          type: string
        criteriaScores:
          type: array
          items:
            $ref: "#/components/schemas/ScreeningCriteriaScore"
      required: [pipelineId, interviewerId]
    SubmitVerdictRequest:
      type: object
      properties:
        verdict:
          type: string
          enum: [shortlist, hold, reject]
        overallScore:
          type: integer
        notes:
          type: string
      required: [verdict]

    # ASSESSMENTS ───────────────────────────────────────────────────────────────
...